requirements.txt
requests==2.31.0
mysql-connector-python==8.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pymysql==1.1.0
pandas==2.1.4 